fix(KTable): apply pr feedback
